# docs/openapi/cloud-server-openapi.yaml
openapi: 3.0.1
info:
  title: cloud-server · 云存储平台 API
  description: |
    ## 统一门户接入（云存储侧）

    本产品接入《云平台统一门户》所需的接口，分为两组：

    | 分组 | 路径 | 鉴权 | 调用方 |
    |---|---|---|---|
    | 门户内网接口 | `/api/internal/**` | 请求头 `X-Internal-Key` | 门户主动调用 |
    | 门户运维接口 | `/api/admin/portal/**` | Bearer JWT + admin 角色 | 运维/联调手工触发 |

    ### 测试前必做：点右上角 Authorize
    1. **internalKey**：填平台侧下发的密钥（对应 `app.portal.internal-key`）。
       未配置时后端会 fail-closed，一律返回 `HTTP 401 {"code":40001}`。
    2. **bearerAuth**：填 `admin / Admin@123` 调 `/api/auth/login` 拿到的 `accessToken`
       （**只填 token 本身，不要带 "Bearer " 前缀**；注意 accessToken 在响应 `data.accessToken`，不是 `data.token`）。

    ### 两个必须知道的约定
    - **`price_cents` / `amount_cents` 单位一律是「分」**，不是元。
    - **`/internal/provision` 的 HTTP 状态码恒为 200**，成败看 body 里的 `code`
      （`0` 成功 / `4003` sku 不存在 / `4004` 库存不足 / `5001` 内部错误）。POST 默认的 201 会让门户判错。
  version: v1
servers:
- url: http://127.0.0.1:3001/api
  description: Generated server url
...
paths:
...
  /admin/portal/reconciliation/report:
    post:
      tags:
      - 门户运维接口
      summary: 手工补报当日账单到门户（对账用）
      description: |
        门户的日对账是「你的账单 vs 门户流水」逐笔比对，但门户**不直连你的库**，
        因此需要把账单推给门户 `POST /api/internal/reconciliation/report`。

        - **幂等**：门户按 `(biz_date, product_code, bill_no)` 唯一，重复上报覆盖金额，可安全重放。
        - **上报范围**：只上报发生过真实资金流水的账单（`status=1` 已入账 / `status=2` 已退款）。
          `status=0` 的待扣账单门户侧没有流水，上报只会制造噪声差异；
          其数量通过返回的 `pendingCount` 告知（可用于发现长期卡住未扣的账单）。
        - 定时任务默认每日 06:00 自动上报前一日，本接口用于**手工补报/测试**。
      operationId: report
      parameters:
      - name: date
        in: query
        description: 账期（ISO 日期，如 2026-10-09）；缺省为昨天
        required: false
        schema:
          type: string
          format: date
        example: 2026-10-09
      responses:
        "200":
          description: 上报结果（accepted 表示门户已接收）
          content:
            application/json:
              example:
                code: 0
                message: ok
                data:
                  bizDate: 2026-10-09
                  billCount: 3
                  pendingCount: 2
                  accepted: true
                  message: ok
      security:
      - bearerAuth: []
  /admin/portal/bills/refund:
    post:
      tags:
      - 门户运维接口
      summary: 退款（幂等，累计不得超原单金额）
      description: |
        1. 以 `type=3`（退款）、`status=0` **先落库**一张退款单（`refund_bill_no` 是退款幂等键）；
        2. 调门户 `POST /wallet/refund`；
        3. 成功后退款单置 `status=1`；**累计退清**时把原单置为 `status=2`（已退款），
           部分退款则原单保持 `status=1`。

        本侧会先做「累计退款不得超过原单金额」的前置校验（门户侧也会拦，返回 `3004`）。

        返回的 `portalCode`：`0` 成功 / `3004` 超退（message 里给出**还能退多少分**）/ `3005` 原单不存在。
      operationId: refund
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRefundRequest"
        required: true
      responses:
        "200":
          description: HTTP 200 恒定，结果看 data.settled / data.portalCode
          content:
            application/json:
              examples:
                退款成功:
                  description: 退款成功
                  value:
                    code: 0
                    message: ok
                    data:
                      refundBillNo: storage-20261009-000005
                      settled: true
                      portalCode: 0
                      message: ok
                累计超退被拦:
                  description: 累计超退被拦
                  value:
                    code: 0
                    message: ok
                    data:
                      refundBillNo: null
                      settled: false
                      portalCode: 3004
                      message: 累计退款超过原单金额，最多还能退 600 分
      security:
      - bearerAuth: []
  /admin/portal/bills/charge:
    post:
      tags:
      - 门户运维接口
      summary: 手工出账并扣费（验证「先落库再扣费」）
      description: |
        完整走一遍 `b_bill` 的扣费流程，用于在**不需要真实按量出账**的情况下验证机制：

        1. 生成账单号（`{product_code}-{yyyyMMdd}-{6位序列}`）并以 `status=0`（待扣）**先落库**；
        2. 调门户 `POST /wallet/deduct`（带 `X-Internal-Key`）；
        3. 按响应码收敛状态：

        | 门户 code | 本侧处理 | 账单状态 |
        |---|---|---|
        | `0` | 成功 | `1` 已入账 |
        | `3003` | **当成功处理**（幂等重复，取首次结果） | `1` 已入账 |
        | `3001` 余额不足 / `3002` 钱包冻结 | **不重试** | 保持 `0` 待扣，记 `fail_code` |
        | 网络超时 / 5xx | **用同一个 bill_no 退避重试**（默认 1s/5s/30s） | 用尽后保持 `0`，`fail_code=5001` |

        ### 怎么验证「先落库再扣费」
        把门户地址指向一个不可达端口，然后调本接口：
        **响应是失败，但 `b_bill` 里一定有一条 `status=0` 的记录**（钱没扣成，账记得住，可对账可补扣）。

        ⚠️ 门户持续系统失败时本接口最坏约 36 秒才返回（退避重试所致）。
        可用环境变量 `PORTAL_DEDUCT_RETRY_DELAYS_MS` 缩短，如 `100,100,100`。
      operationId: charge
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminChargeRequest"
        required: true
      responses:
        "200":
          description: HTTP 200 恒定，结果看 data.settled / data.portalCode
          content:
            application/json:
              examples:
                扣费成功:
                  description: 扣费成功
                  value:
                    code: 0
                    message: ok
                    data:
                      billNo: storage-20261009-000002
                      settled: true
                      portalCode: 0
                      message: ok
                      retryCount: 0
                余额不足（不重试）:
                  description: 余额不足（不重试）
                  value:
                    code: 0
                    message: ok
                    data:
                      billNo: storage-20261009-000004
                      settled: false
                      portalCode: 3001
                      message: 余额不足
                      retryCount: 0
                门户不可达（已重试3次，账单仍落库）:
                  description: 门户不可达（已重试3次，账单仍落库）
                  value:
                    code: 0
                    message: ok
                    data:
                      billNo: storage-20261009-000001
                      settled: false
                      portalCode: 5001
                      message: 扣费系统失败，已重试 3 次：...
                      retryCount: 3
      security:
      - bearerAuth: []
...
  /internal/products:
    get:
      tags:
      - 门户内网接口
      summary: ① 商品目录（门户每 10 分钟拉取）
      description: |
        返回本产品**全部在售** SKU，门户据此全量 upsert 到它自己的 `p_product` 缓存表。

        - **无分页**：一次全量返回，门户超时 10 秒，故本接口不做重计算。
        - **`price_cents` 单位是分**（不是元）。
        - 不再上报的 sku 会被门户置 `is_active=0` 软下架，因此 **`sku_id` 必须稳定**：
          改 sku_id 等于下架旧的、上架新的。
        - `spec` 结构完全由本产品自定，门户原样展示（门户不理解你的字段）。
        - `stock` 为 `null` 表示不限库存（云存储即如此），前端显示「现货」。
      operationId: products
      responses:
        "200":
          description: 成功（注意：业务失败也返回 HTTP 200，靠 body.code 区分）
          content:
            application/json:
              examples:
                "4 个种子 SKU":
                  description: 4 个种子 SKU
                  value:
                    code: 0
                    message: ok
                    data:
                      products:
                      - sku_id: storage-100g
                        name: 云存储·100G
                        spec:
                          quota_gb: 100
                          transfer: unlimited
                          period: 30天
                        price_cents: 10000
                        billing_mode: 1
                        period_days: 30
                        stock: null
                      - sku_id: storage-500g
                        name: 云存储·500G
                        spec:
                          quota_gb: 500
                          transfer: unlimited
                          period: 30天
                        price_cents: 50000
                        billing_mode: 1
                        period_days: 30
                        stock: null
        "401":
          description: 内部密钥缺失或错误（未配置 app.portal.internal-key 时同样拒绝）
          content:
            application/json:
              example:
                code: 40001
                message: 无效的内部调用凭证
                data: null
      security:
      - internalKey: []
...
  /admin/portal/status:
    get:
      tags:
      - 门户运维接口
      summary: 门户接入状态速览（排障）
      description: |
        仅返回一个提示串，**不返回任何密钥明文**。真正的连通性请用：
        - `GET /api/internal/products`（验证 `X-Internal-Key` 与门户拉取链路）
        - `POST /api/admin/portal/reconciliation/report`（验证本侧 → 门户的出站链路）
      operationId: status_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultMapStringObject"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    ResultMapStringObject:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: object
          additionalProperties:
            type: object
...
    AdminRefundRequest:
      required:
      - amountCents
      - billNo
      type: object
      properties:
        billNo:
          type: string
          description: 原扣费账单号（`b_bill.bill_no`，形态 `storage-20261009-000002`）
          example: storage-20261009-000002
        amountCents:
          minimum: 1
          type: integer
          description: 本次退款金额，单位**分**（必须 > 0，且累计退款不得超过原单金额）
          format: int64
          example: 1000
        reason:
          type: string
          description: 退款原因（可空，缺省为「管理端发起退款」）
          example: 开通失败
      description: 退款请求（注意：本请求体字段是 camelCase）
    AdminChargeRequest:
      required:
      - amountCents
      - portalUserId
      - type
      type: object
      properties:
        portalUserId:
          type: string
          description: 门户用户 UUID（门户 JWT 的 sub）；本侧据此解析/建立影子本地用户
          example: 018f3a2b-7c1d-7000-9e21-3f9a2b8c1d00
        amountCents:
          minimum: 1
          type: integer
          description: 出账金额，单位**分**（必须 > 0）
          format: int64
          example: 3000
        type:
          type: integer
          description: 账单类型：`1` 续费 / `2` 按量（`3` 退款由退款接口生成，此处不可传）
          format: int32
          example: 2
          enum:
          - 1
          - 2
        relatedBillNo:
          type: string
          description: 关联业务单号（可空），仅用于溯源
          example: verify-001
      description: 手工出账请求（注意：本请求体字段是 camelCase）
...
  securitySchemes:
    internalKey:
      type: apiKey
      description: 门户 → 本产品的内部调用密钥（后端配置项 app.portal.internal-key）。仅作用于 /internal/**，未配置该密钥时后端拒绝一切内部调用（fail-closed）。
      name: X-Internal-Key
      in: header
    bearerAuth:
      type: http
      description: 两种令牌都被接受：① 本系统自签 HS256（POST /api/auth/login 获取，管理端接口用这个）；② 门户签发的
        RS256（业务路由用，后端按 JWKS 本地验签）。
      scheme: bearer
      bearerFormat: JWT